POM file that builds the output NBM.
Adds the POM file that uses nbm-maven-plugin to build the output NBM
file. The changes - besides adding POM - are minimal, just to make the
build successful and produce a working NBM file.

The output NBM was tested against JVisualVM included in the latest JDK 7
(1.7.0_80-b15), JDK 8 (1.8.0_181-b13) and the standalone JVisualVM
1.4.1.